Fix: Gastos, compras y cantidades sin decimales
Cambios realizados:
- Gastos: Agregado campo 'color' a ExpenseCategory
- Gastos: Agregado selector de metodo de pago en formulario
- Compras: Cantidades ahora son enteros (PositiveIntegerField)
- Stock: current_stock, min_stock, max_stock son enteros
- Stock: StockMovement.quantity es entero
- POS: POSTransactionItem.quantity es entero
- Formularios actualizados con step=1 para cantidades
- Agregado script de pruebas funcionales

+++ sales/models.py
@@ -91,10 +91,8 @@ class SaleItem(models.Model):
         'Nombre del Producto',
         max_length=200
     )
-    quantity = models.DecimalField(
-        'Cantidad',
-        max_digits=10,
-        decimal_places=3
+    quantity = models.PositiveIntegerField(
+        'Cantidad'
     )
     unit_price = models.DecimalField(
         'Precio Unitario',
